requirements.txt
Django==4.2.9
djangorestframework==3.14.0
gunicorn==21.2.0
whitenoise==6.6.0
python-dateutil==2.8.2
